+++ test/correctness/PySys_internal_178/Input/MyNestedTestcase/pysystest.py
@@ -1,7 +1,7 @@
 __pysys_title__   = r""" Nested """ 
 #                        ================================================================================
 
-__pysys_modes__ = "lambda helper: {'MyMode': {}}"
+__pysys_modes__ = "lambda helper: {'MyMode': {'ModeParamA':123, 'ModeParamB':'Foo'}}"
 
 import pysys
 from pysys.constants import *